Remove GetDescriptorInfo and store in Buffer object. (#358)
This Cl stores the relevant data into the amber::Buffer and removes the
need to call GetDescriptorInfo to retrieve the data.

+++ src/vulkan/vertex_buffer_test.cc
@@ -32,7 +32,7 @@ const VkPhysicalDeviceMemoryProperties kMemoryProperties =
 class BufferForTest : public Buffer {
  public:
   BufferForTest(Device* device,
-                size_t size_in_bytes,
+                uint32_t size_in_bytes,
                 const VkPhysicalDeviceMemoryProperties& properties)
       : Buffer(device, size_in_bytes, properties) {
     memory_.resize(4096);
